refactor: :memo: renames poo not required file name to match component name
